[Maintenance] Add tests for all supported datasets (#42)
Add test for verification of runner_config.py
Add pytest to precommit
Added test for Coronahack pipeline
Added dummy data for Coronahack pipeline
Created library for pipeline testing
throw error if source path exists but empty label path or masks_path 
throw error if X is empty
check if an image already has a label
clean temp PNG files
Add KITS23 Dummy data
Add Stanford Coca dummy data

---------

+++ src/steps/delete_imgs_with_no_annotations.py
@@ -68,4 +68,5 @@ class DeleteImgsWithNoAnnotations(TransformerMixin):
                 if no_label:
                     # If there is a blank mask and no label
                     os.remove(img_path)
-                    os.remove(mask_path)
+                    if os.path.exists(mask_path):
+                        os.remove(mask_path)
